requirements.txt
pystray==0.19.0
Pillow==10.4.0
pyinstaller
pytest==8.2.2
pywin32
keyboard==0.13.5
requests==2.32.3
PyQt6==6.7.1
